+++ codec/legacy/codec.go
@@ -1,0 +1,20 @@
+package legacy
+
+import (
+	"github.com/cosmos/cosmos-sdk/codec"
+	cryptocodec "github.com/cosmos/cosmos-sdk/crypto/codec"
+)
+
+//nolint:stylecheck // ignore linting of comment
+// Deprecated: Cdc defines a global generic sealed Amino codec to be used throughout sdk. It
+// has all Tendermint crypto and evidence types registered.
+//
+// TODO: remove this global.
+var Cdc *codec.Codec
+
+func init() {
+	Cdc = codec.New()
+	cryptocodec.RegisterCrypto(Cdc)
+	codec.RegisterEvidences(Cdc)
+	Cdc.Seal()
+}